Refactoring scrill area input WIP

+++ o2Editor/Sources/Core/EditorApplication.cpp
@@ -223,7 +223,7 @@ namespace Editor
 		if (o2Input.IsKeyDown(VK_F6)) 
 		{
 			for (int i = 0; i < mActions.Count(); i++)
-				o2Debug.DrawText(Vec2F(0, 20 * i), (String)i + mActions[i]->GetName());
+				o2Debug.DrawText(Vec2F(0, (float)(20 * i)), (String)i + mActions[i]->GetName());
 		}
 	}
